System.IO.Streams.Debug

Description

Convenience module for debugging streams. Provides stream transformers that wrap InputStreams and OutputStreams, sending a description of all data to an OutputStream for debugging.

Debuggers

debugInput

Arguments

:: (a -> ByteString)
function to convert stream elements to ByteString

-> ByteString
name of this debug stream, will be prepended to debug output

-> OutputStream ByteString
stream the debug info will be sent to

-> InputStream a
input stream

-> IO (InputStream a)

debugOutput

Arguments

:: (a -> ByteString)
function to convert stream elements to ByteString

-> ByteString
name of this debug stream, will be prepended to debug output

-> OutputStream ByteString
debug stream

-> OutputStream a
output stream

-> IO (OutputStream a)

debugInputBS

Arguments

:: ByteString
name of this debug stream, will be prepended to debug output

-> OutputStream ByteString
stream the debug info will be sent to

-> InputStream ByteString
input stream

-> IO (InputStream ByteString)

debugOutputBS

Arguments

:: ByteString
name of this debug stream, will be prepended to debug output

-> OutputStream ByteString
stream the debug info will be sent to

-> OutputStream ByteString
output stream

-> IO (OutputStream ByteString)