System.IO.Streams.Network

Description

Converting network Sockets to streams.

Sockets to Streams

socketToStreams :: Socket -> IO (InputStream ByteString, OutputStream ByteString)

Converts a Socket to an InputStream / OutputStream pair. Note that, as is usually the case in io-streams, writing a Nothing to the generated OutputStream does not cause the underlying Socket to be closed.

socketToStreamsWithBufferSize

Arguments

:: Int
how large the receive buffer should be

-> Socket
network socket

-> IO (InputStream ByteString, OutputStream ByteString)

Converts a Socket to an InputStream / OutputStream pair, with control over the size of the receive buffers. Note that, as is usually the case in io-streams, writing a Nothing to the generated OutputStream does not cause the underlying Socket to be closed.