System.IO.Streams.Text

Description

Stream primitives for decoding and encoding Text values in UTF-8 format.

Decoders and Encoders

decodeUtf8 :: InputStream ByteString -> IO (InputStream Text)

Decode an InputStream of ByteStrings in UTF-8 format into an InputStream of Text values. If decoding fails, will throw an exception. See Data.Text.Encoding.decodeUtf8.

decodeUtf8With :: OnDecodeError -> InputStream ByteString -> IO (InputStream Text)

Decode an InputStream of ByteStrings in UTF-8 format into an InputStream of Text values. If decoding fails, invokes the given OnDecodeError function to decide what to do. See Data.Text.Encoding.decodeUtf8With.

encodeUtf8 :: OutputStream ByteString -> IO (OutputStream Text)

Convert an OutputStream taking ByteStrings to an OutputStream that takes Text, encoding the data as UTF-8. See Data.Text.Encoding.encodeUtf8.